switch UserInfection to Fragment
